Rework lib to share code and support more devices
