+++ nodes/miio-humidifier-in.html
@@ -18,10 +18,10 @@
         },
         inputs: 0,
         outputs: 1,
-        outputLabels: ["homekit"],
-        paletteLabel: 'in',
+        outputLabels: ["homekit + raw"],
+        paletteLabel: 'Humidifier In',
         label: function () {
-            return this.name || "Air Humidifier In";
+            return this.name || "Humidifier In";
         }
     });
 </script>
